requirements.txt
openai>=1.3.9
chromadb>=0.4.24
python-dotenv>=1.0.1